uadk: fix truncation before type conversion
Truncation may occur before type conversion,
use force cast to eliminate errors.

+++ v1/wd_rsa.c
@@ -178,7 +178,7 @@ static int kg_in_param_check(void *ctx, struct wd_dtb *e,
 		return -WD_EINVAL;
 	}
 
-	kg_in_size = GEN_PARAMS_SZ(c->key_size);
+	kg_in_size = (int)GEN_PARAMS_SZ(c->key_size);
 	if (unlikely(br->get_bufsize &&
 	    br->get_bufsize(br->usr) < (kg_in_size + sizeof(*kg_in)))) {
 		WD_ERR("Blk_size < need_size<0x%lx>.\n", (kg_in_size + sizeof(*kg_in)));
@@ -202,7 +202,7 @@ struct wcrypto_rsa_kg_in *wcrypto_new_kg_in(void *ctx, struct wd_dtb *e,
 		return NULL;
 
 	br = &c->setup.br;
-	kg_in_size = GEN_PARAMS_SZ(c->key_size);
+	kg_in_size = (int)GEN_PARAMS_SZ(c->key_size);
 	kg_in = br->alloc(br->usr, kg_in_size + sizeof(*kg_in));
 	if (unlikely(!kg_in)) {
 		WD_ERR("ctx br->alloc kg_in memory fail!\n");
@@ -285,7 +285,7 @@ struct wcrypto_rsa_kg_out *wcrypto_new_kg_out(void *ctx)
 	if (c->setup.is_crt)
 		kg_out_size = CRT_GEN_PARAMS_SZ(c->key_size);
 	else
-		kg_out_size = GEN_PARAMS_SZ(c->key_size);
+		kg_out_size = (int)GEN_PARAMS_SZ(c->key_size);
 
 	br = &c->setup.br;
 	if (unlikely(!br->alloc)) {
@@ -434,7 +434,7 @@ static int create_ctx_key(struct wcrypto_rsa_ctx_setup *setup,
 
 	if (setup->is_crt) {
 		len = sizeof(struct wcrypto_rsa_prikey) +
-			CRT_PARAMS_SZ(ctx->key_size);
+			(__u32)CRT_PARAMS_SZ(ctx->key_size);
 		if (br->get_bufsize && br->get_bufsize(br->usr) < len) {
 			WD_ERR("Blk_size < need_size<0x%x>.\n", len);
 			return -WD_ENOMEM;
@@ -449,7 +449,7 @@ static int create_ctx_key(struct wcrypto_rsa_ctx_setup *setup,
 		init_pkey2(pkey2, ctx->key_size);
 	} else {
 		len = sizeof(struct wcrypto_rsa_prikey) +
-			GEN_PARAMS_SZ(ctx->key_size);
+			(__u32)GEN_PARAMS_SZ(ctx->key_size);
 		if (br->get_bufsize && br->get_bufsize(br->usr) < len) {
 			WD_ERR("Blk_size < need_size<0x%x>.\n", len);
 			return -WD_ENOMEM;
@@ -464,7 +464,7 @@ static int create_ctx_key(struct wcrypto_rsa_ctx_setup *setup,
 		init_pkey1(pkey1, ctx->key_size);
 	}
 
-	len = sizeof(struct wcrypto_rsa_pubkey) + GEN_PARAMS_SZ(ctx->key_size);
+	len = sizeof(struct wcrypto_rsa_pubkey) + (__u32)GEN_PARAMS_SZ(ctx->key_size);
 	if (br->get_bufsize && br->get_bufsize(br->usr) < len) {
 		br->free(br->usr, ctx->prikey);
 		WD_ERR("Blk_size < need_size<0x%x>.\n", len);
